Added support for Windows and WinAPI

+++ include/thread.h
@@ -3,5 +3,5 @@
 #include <stdatomic.h>
 #include <core.h>
 
-typedef int (*sr_thread_t)(void *);
+typedef unsigned long (*sr_thread_t)(void *);
 void sr_thread_create(sr_thread_t thread_func, void * thread_args);
